Added animation to flip a card and information is on the back along with the option to delete.

+++ smurfs/src/components/App.js
@@ -34,7 +34,7 @@ class App extends Component {
   }
   deleteSmurf = id => {
     const sureToDelete = prompt('This can not be undone. To continue enter the word delete (just the word no extra characters)')
-    if(sureToDelete.toLowerCase() === 'delete'){
+    if(sureToDelete && sureToDelete.toLowerCase() === 'delete'){
       this.props.deleteSmurf(id); 
     }
   }
@@ -43,26 +43,35 @@ class App extends Component {
     this.resetFields()
     this.props.updateSmurf(this.props.current, smurf)
   }
+  flipCard = index => {
+    const cards = document.querySelectorAll('.card')
+    const card = cards[index]
+    card.classList.toggle('is-flipped')
+  }
 
   render() {
     const smurfs = this.props.smurfs; 
     return (
       <div className="App">
-        <h1>SMURFS! 2.0 W/ Redux</h1>
+        <div className = "header"><h1>SMURFS! 2.0 W/ Redux</h1></div>
+        <div><p>Instructions: click to flip card for more infomation on the smurf. Double click to update. Click icon on back side of card to delete(must flip first).</p></div>
+        <div className = "form">
         <input type="text" placeholder="...name" ref ={input => this.name = input}/>
         <input type="text" placeholder="...age" ref ={input => this.age = input}/>
         <input type="text" placeholder ="...height" ref ={input => this.height = input}/>
         {this.props.doubleClicked === false ? <button onClick = {this.addingSmurf}>Add Smurf</button> 
         : 
         <button onClick ={this.updating}>Update Smurf</button>
-        }        
+        } 
+        </div>      
+        <br/>
         <br/>
         <br/>
         <br/>
         <br/>
         {smurfs.map((smurf, i )=> 
-        <div onDoubleClick = {() => this.doubleClickingSmurf(smurf.id, smurf.name, smurf.age, smurf.height)}key ={i} className = 'scene'>
-          <div className ="card">
+        <div onClick = {() => this.flipCard(i)} onDoubleClick = {() => this.doubleClickingSmurf(smurf.id, smurf.name, smurf.age, smurf.height)}key ={i} className = 'scene'>
+          <div  className ="card">
             <div className="card__face card__face--front">
             {smurf.name}
             </div> 
@@ -77,6 +86,7 @@ class App extends Component {
             </div>
           </div>
         </div>)}
+
       </div>
     );
   }
